finish improvements
